requirements.txt
numpy
pandas
torch
scikit-learn
scipy
matplotlib
openpyxl
perceval-quandela
merlin-python